pkg/codebase: do not split codebases graphql
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 7cc18852-5245-47e1-a119-d35364ed3156
workspace_id: 002baaf8-41ad-45ce-bfec-b63cee5d6cac
