+++ pkg/codebase/graphql/module.go
@@ -1,0 +1,9 @@
+package graphql
+
+import (
+	"mash/pkg/di"
+)
+
+var Module = di.NewModule(
+	di.ProvidesCycle(NewCodebaseRootResolver),
+)
