+++ pkg/onboarding_test.go
@@ -27,7 +27,7 @@ import (
 	workers_ci "mash/pkg/ci/workers"
 	"mash/pkg/codebase"
 	db_codebase "mash/pkg/codebase/db"
-	graphql_codebase "mash/pkg/codebase/graphql/oss"
+	graphql_codebase "mash/pkg/codebase/graphql"
 	routes_v3_codebase "mash/pkg/codebase/routes"
 	service_codebase "mash/pkg/codebase/service"
 	db_comments "mash/pkg/comments/db"
@@ -330,6 +330,7 @@ func TestCreate(t *testing.T) {
 		nil,
 		nil,
 		&changeRootResolver,
+		nil,
 		nil,
 		nil, // instantIntegrationRootResolver
 		logger,
